requirements.txt
amqp==5.1.1
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.11.0
django-debug-toolbar==3.5.0
django-extensions==3.1.5
django-filter==21.1
django-treebeard==4.5.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
executing==0.8.3
fake-factory==9999.9.9
Faker==13.11.0
idna==3.3
importlib-metadata==4.11.3
ipython==8.2.0
jedi==0.18.1
jsonfield==3.1.0
kombu==5.2.4
Markdown==3.3.6
matplotlib-inline==0.1.3
netaddr==0.8.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
PyJWT==2.3.0
pyllist==0.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
urllib3==1.26.9
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
